+++ sme.gclass.api.infra/Dtos/Sga/FiltroConsultaSgaFuncionarioProfessorDto.cs
@@ -1,4 +1,4 @@
-﻿namespace SME.GoogleClassroom.Worker.Rabbit.Controllers
+﻿namespace SME.GoogleClassroom.Infra.Dtos.Gsa
 {
     public class FiltroConsultaSgaFuncionarioProfessorDto
     {
