+++ sme.gclass.api.worker.rabbit/Controllers/ProfessorController.cs
@@ -194,20 +194,5 @@ namespace SME.GoogleClassroom.Worker.Rabbit.Controllers
             await useCase.Executar();
             return Ok();
         }
-
-        /// <summary>
-        /// Retorna os professores e funcionarios de uma Escola pelo Código da escola e Ano Letivo
-        /// </summary>
-        /// <response code="200">A consulta foi realizada com sucesso.</response>
-        /// <response code="500">Ocorreu um erro inesperado durante a consulta.</response>
-        /// <response code="601">Houve uma falha de validação durante a consulta.</response>
-        [HttpGet("funcionarios/escola/{codigoEscola}/anoLetivo/{anoLetivo}")]
-        [ProducesResponseType(typeof(ProfessoresFuncionariosSgaDto), 200)]
-        [ProducesResponseType(typeof(RetornoBaseDto), 500)]
-        [ProducesResponseType(typeof(RetornoBaseDto), 601)]
-        public async Task<IActionResult> ObterProfessoresEFuncionariosSga(string codigoEscola,int anoLetivo, [FromServices] IFuncionariosProfessoresEolSgaUseCase useCase)
-        {
-            return Ok(await useCase.Executar(anoLetivo,codigoEscola));
-        }
     }
 }
